move intake commands
